requirements.txt
# ETL & Pipeline Requirements
pandas
numpy
thefuzz
python-levenshtein
psycopg2-binary
python-dotenv
requests
undetected-chromedriver
selenium
beautifulsoup4
playwright
playwright-stealth
boto3
google-generativeai
